requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6
PyMuPDF>=1.23.0
pytesseract>=0.3.10
Pillow>=10.2.0
ollama>=0.4.0
groq>=0.4.0
python-dotenv>=1.0.0
pydantic>=2.5.0
